refactor(domain): move entryModel under models folder
