+++ frontend/src/Domain/Interfaces/Entries/EntryRepositoryInterface.ts
@@ -11,7 +11,7 @@
  * - No Application DTOs here; only domain types.
  * - `save` performs upsert semantics.
  */
-import type {Entry} from '@src/Domain/Entries/Entry';
+import type {Entry} from '@src/Domain/Models/Entries/Entry';
 import type {ListEntriesCriteriaInterface} from '@src/Domain/Interfaces/Entries/ListEntriesCriteriaInterface';
 import type {ListEntriesPageInterface} from '@src/Domain/Interfaces/Entries/ListEntriesPageInterface';
