+++ frontend/src/Domain/Models/Entries/Entry.ts
